requirements.txt
absl-py==2.1.0
amqp==5.2.0
astunparse==1.6.3
async-timeout==4.0.3
billiard==4.2.0
blinker==1.6.2
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.5
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.6
dnspython==2.6.1
eventlet==0.36.1
filelock==3.12.0
flask==2.3.2
flatbuffers==24.3.25
fonttools==4.46.0
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.1.2
jinja2==3.1.2
keras==3.3.3
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
lightgbm==4.0.0
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.0.2
pillow==10.1.0
platformdirs==3.5.1
prompt-toolkit==3.0.43
protobuf==4.25.3
pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
redis==5.0.4
requests==2.31.0
rich==13.7.1
scipy==1.11.1
setuptools==69.5.1
six==1.16.0
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf-keras==2.16.0
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing-extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
vine==5.1.0
virtualenv==20.23.0
wcwidth==0.2.13
werkzeug==2.3.6
wheel==0.43.0
wrapt==1.16.0
xgboost==1.7.6
